filters: Fix has:image and avoid future issues for other has filters.
This commit solves a bug due non working find() for a html element
which was not a descendant and fixes it by wrapping a virtual div
around the message content so that find works as intended in any
case whatsoever. This solution even protects from future cases of
html content or hierarchy changes.

Fixes #16118.